pyproject.toml
[tool.poetry]
name = "cybersec-advisory-system"
version = "0.1.0"
description = "Multi-Agent Cybersecurity Advisory System with LangGraph, Langfuse, and MCP integration"
authors = ["Daniel Andersson <daniel.andersson311@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<4.0"

# Core Configuration
pydantic = ">=2.10.0"
pydantic-settings = ">=2.6.0"
python-dotenv = ">=1.0.0"

# LLM & AI Framework
langchain = ">=0.3.0"
langchain-openai = ">=0.3.0"
langgraph = ">=0.3.27"
openai = ">=1.50.0"

# Observability
langfuse = ">=3.0.0"

# Database
psycopg2-binary = ">=2.9.0"
asyncpg = ">=0.29.0"

# Vector Database
qdrant-client = ">=1.7.0"

# Web Framework
fastapi = ">=0.100.0"
uvicorn = ">=0.26.0"

# Utilities
aiofiles = ">=24.0.0"
fastmcp = "^2.10.6"
tavily-python = "^0.7.10"
langchain-qdrant = "^0.2.0"
langchain-community = "^0.3.27"
sentence-transformers = "^5.0.0"
pypdf = "^5.9.0"
pytest-asyncio = "^1.1.0"
pytest = "^8.4.1"
instructor = "^1.10.0"
async-lru = "^2.0.5"
python-docx = "^1.2.0"
unstructured = {extras = ["md"], version = "^0.18.11"}
fastembed = "^0.7.1"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
black = ">=24.0.0"
ruff = ">=0.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"